重构了切换表盘，DIY表盘界面的代码框架，优化了找手机以及远程拍照的响应速度

+++ app/src/main/java/com/szip/sportwatch/Util/FileUtil.java
@@ -110,8 +110,11 @@ public class FileUtil {
     }
 
     public void deleteFile(String fileName){
-        File file = new File(fileName);
-        file.delete();
+        if(fileName!=null) {
+            File file = new File(fileName);
+            if (file != null)
+                file.delete();
+        }
     }
 
     public String getPath() {
